requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
pydantic==2.5.0
numpy<2.0.0
scipy<1.14.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.9
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
bcrypt==4.1.2
pydantic[email]
fpdf2==2.8.1
supabase
alembic==1.13.1
